Enum raqote::Source

pub enum Source<'a> {
    Solid(SolidSource),
    Image(Image<'a>, ExtendMode, FilterMode, Transform),
    RadialGradient(Gradient, Spread, Transform),
    TwoCircleRadialGradient(Gradient, Spread, Point, f32, Point, f32, Transform),
    LinearGradient(Gradient, Spread, Transform),
}

LinearGradients have an implicit start point at 0,0 and an end point at 256,0. The transform parameter can be used to adjust them to the desired location. RadialGradients have an implict center at 0,0 and a radius of 128. The helper functions: new_linear_gradient, new_radial_gradient and new_two_circle_radial_gradient allow the gradients to be constructed with easier to understand inputs. The transform parameter maps user space to source space. This means that setting the same transform on the draw target as the source will have the effect of canceling out.

These locations are an artifact of the blitter implementation and will probably change in the future to become more ergonomic.

Variants

Solid

Tuple Fields of Solid

0: SolidSource

Image

Tuple Fields of Image

0: Image<'a>1: ExtendMode2: FilterMode3: Transform

RadialGradient

Tuple Fields of RadialGradient

0: Gradient1: Spread2: Transform

TwoCircleRadialGradient

Tuple Fields of TwoCircleRadialGradient

0: Gradient1: Spread2: Point3: f324: Point5: f326: Transform

LinearGradient

Tuple Fields of LinearGradient

0: Gradient1: Spread2: Transform

Implementations

impl<'a> Source<'a>

pub fn new_linear_gradient(
gradient: Gradient,
start: Point,
end: Point,
spread: Spread
) -> Source<'a>

Creates a new linear gradient source where the start point corresponds to the gradient stop at position = 0 and the end point corresponds to the gradient stop at position = 1.

pub fn new_radial_gradient(
gradient: Gradient,
center: Point,
radius: f32,
spread: Spread
) -> Source<'a>

Creates a new radial gradient that is centered at the given point and has the given radius.

pub fn new_two_circle_radial_gradient(
gradient: Gradient,
center1: Point,
radius1: f32,
center2: Point,
radius2: f32,
spread: Spread
) -> Source<'a>

Creates a new radial gradient that is centered at the given point and has the given radius.